requirements.txt
PyPDF2==3.0.1
openpyxl==3.1.2
Flask==3.0.0
Werkzeug==3.0.1
gunicorn==21.2.0
groq==0.11.0
python-dotenv==1.0.0
